requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
gunicorn==0.17.2
Flask-SQLAlchemy==0.16
sqlalchemy-migrate==0.7.2
WTForms==1.0.4
